chore: Move header validation into `sync::headers::validation`
