+++ dash-spv/src/client/sync_coordinator.rs
@@ -17,8 +17,8 @@ use crate::error::{Result, SpvError};
 use crate::network::constants::MESSAGE_RECEIVE_TIMEOUT;
 use crate::network::NetworkManager;
 use crate::storage::StorageManager;
+use crate::sync::headers::validate_headers;
 use crate::types::{DetailedSyncProgress, SyncProgress};
-use crate::validation::validate_headers;
 use crate::ValidationMode;
 use key_wallet_manager::wallet_interface::WalletInterface;
 use std::time::{Duration, Instant, SystemTime};
