+++ dash-spv/src/sync/headers/mod.rs
@@ -1,5 +1,7 @@
 //! Header synchronization with fork detection and reorganization handling.
 
 mod manager;
+pub mod validation;
 
 pub use manager::{HeaderSyncManager, ReorgConfig};
+pub use validation::validate_headers;
