+++ dash-spv/src/validation/mod.rs
@@ -1,9 +1,7 @@
 //! Validation functionality for the Dash SPV client.
 
-pub mod headers;
 pub mod instantlock;
 pub mod quorum;
 
-pub use headers::validate_headers;
 pub use instantlock::InstantLockValidator;
 pub use quorum::{QuorumInfo, QuorumManager, QuorumType};
